Renaming solitaireEnhancer to enhanceSolitaire. Moving is game complete logic to SolitaireEnhancer
