+++ src/lib/enhancers.ts
@@ -2,10 +2,10 @@ import { LocationAwarePotentiallyUndefinedSolitaireCard, Solitaire } from "types
 import { CardEnhancer } from "./enhancers/cardEnhancer";
 import { SolitaireEnhancer } from "./enhancers/solitaireEnhancer";
 
-export const enhanceCard = (card: LocationAwarePotentiallyUndefinedSolitaireCard) => {
+export const enhanceCard = (card: LocationAwarePotentiallyUndefinedSolitaireCard): CardEnhancer => {
     return new CardEnhancer(card);
 };
 
-export const solitaireEnhancer = (solitaire: Solitaire) => {
+export const enhanceSolitaire = (solitaire: Solitaire): SolitaireEnhancer => {
     return new SolitaireEnhancer(solitaire);
 }
